\name{RMesquite-package}
\alias{RMesquite-package}
\alias{RMesquite}
\docType{package}
\title{
RMesquite - Provides a transparent wrapper for Mesquite functions in R
}
\description{
  The RMesquite package provides transparent access to data types and
  analysis modules in Mesquite. Mesquite is a modular system for
  evolutionary analysis in Java. Mesquite is available from
  http://mesquiteproject.org.
}
\details{
\tabular{ll}{
Package: \tab RMesquite\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2007-12-11\cr
License: \tab GPLv2.0 or later\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Hilmar Lapp (hlapp@gmx.net) and Wayne Maddison (wmaddisn@interchange.ubc.ca)

Maintainer: Hilmar Lapp <hlapp@gmx.net>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
